Added getAttributesAndReferences function to dashboard content management service
